refactor(rust-port): start pi-natives core/facade split with pi-natives-core
Establishes the roadmap phase 1 pattern: algorithmic cores move to the
new NAPI-free pi-natives-core crate while pi-natives keeps thin facades
with unchanged NAPI export names and signatures (Bun runtime on dev is
unaffected; cargo check -p pi-natives stays green).

Migrated in this step: glob_util (string errors in core, NAPI error
conversion in the facade), hashline (UTF-16 core with byte-identical
TS parity vectors from hashline/hash.ts; JsString/Utf16String interop
stays in the facade), linediff (jsdiff-parity algorithm in core, NAPI
object re-wrap in the facade).

crates/gjc now depends on pi-natives-core; cargo tree confirms napi is
absent from both the gjc and pi-natives-core dependency trees. The
remaining modules migrate module-by-module (tracked in the ultragoal
ledger and roadmap).
